moved frontend trip components into a "trip" directory
