+++ frontend/components/Trip/Trip.tsx
