+++ frontend/components/Trip/TripMap.tsx
